+++ DADback-end/DADback-end/Model/Disciplinas.cs
@@ -8,7 +8,7 @@ namespace DADback_end.Model
 {
     public class Disciplinas
 {
-        public long? id { get; set; }
+        public int? id { get; set; }
         public string nome { get; set; }
         
 }
